requirements.txt
sounddevice==0.5.2
PyAudio==0.2.14
loguru==0.7.3
piper-tts==1.3.0
scipy==1.16.2
av==15.1.0
websocket-client==1.8.0
https://github.com/abetlen/llama-cpp-python/releases/download/v0.3.2/llama_cpp_python-0.3.2-cp311-cp311-linux_x86_64.whl
